add cpuidle_02 and cpuidle_03 tests
This patch provides a couple of tests for cpuidle in addition with
a small program which behaves to sollicitate the cpuidle drivers
especially when both cores needs to reach the same C-states.

The program forks as many online processors present on the system
and set the affinity to it. Then it runs, during 120 secs, small
chunks of busy loops and idle loops. When there is something wrong
with the cpuidle driver, this program often triggers the problem which
results in a kernel panic, tasks hung warning or system hang.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>