thermal: add switch for thermal_06.sh
This script aims at testing all the trip points are crossed or not, since the
trip points are usually very high and can be reached only in extreme conditions
like high environment temerature etc, in normal conditions they cannot be all
crossed even if the CPU load is high, so failures will be reported and this
will introduce a mess.
Add this switch to switch off this test by default, anyone who wants to do this
test still can enable it manually.

Signed-off-by: hongbo.zhang <hongbo.zhang@linaro.com>