thermal: don't allow the thermal governor to change cdev states in thermal_02
When running thermal_02, the thermal governor can decide to change the
cooling device state between setting the state and reading the state,
so this test may fail.

To prevent it, switch all the thermal zones to the user_space thermal
governor before running the test.  This governor never touches cooling
devices.

Signed-off-by: Javi Merino <javi.merino@arm.com>
Signed-off-by: Sanjay Singh Rawat <sanjay.rawat@linaro.org>