cpufreq: fixed array initializations to calculate freq deviation
Resolve 'err' messages by making changes to
calculate sum and average of frequency measurements
in cpufreq_06 test.

In the function 'compute_freq_ratio' the frequency
is stored in an array of variables by indirect reference
(http://www.tldp.org/LDP/abs/html/ivr.html)
but the variable 'index' that points to next element
in the array is initialized to 0 every time
hence overriding the values. The same happens for the
function 'compute_freq_ratio_sum' where the 'index'
variable is always initialized to 0 and the variable
'sum' is always set to 0 not adding the subsequent
values. Hence these initializations must be deleted
from those functions and written into the
'function check_deviation' for keeping
the correct values of frequency and calculations of
average and summatory.

With these changes the tests can now calculate
the real values of average for the frequencies
and the deviations can be tested and the tests
now passes.

This patch fixes this bug:
https://bugs.linaro.org/show_bug.cgi?id=2232

Signed-off-by: Saul Romero <saul.romero@arm.com>
[lisa: Edit parts of commit message.]
Signed-off-by: Lisa Nguyen <lisa.nguyen@linaro.org>